Greatly speed up all-pairs-shortest paths by reducing allocs.
This commit reuses the cache and modifies it on each iteration of Floyd Warshall, instead of
allocation a brand new gigantic vec-of-vecs each time. I was originally worried that updating
in-place might cause issues, but it does not. Note in particular that if we end up with from_vertex
or to_vertex equalling intermediate, nothing bad happens (if there are no negative cycles, the step
does nothing; if there are, we end up potentially using the negative cycle to reduce the cost, but
that's okay since the algorithm will still eventually terminate and we can check at the end for those
cycles). And if intermediate is neither, then updating cache[from][to] won't help any subsequent
from, to pairs with this intermediate.